Fixing Axis2 131
The exploded directory only allowed for services not for modules , and the support for exploded directory handle by the parameter in axis2.xml, if and only if its value is true exploded directory will be supported.  
The parameter look like follows;
<parameter name="extarct_service_archieve" locked="xsd:false">true</parameter>

if the value is true , first extract all the archive file in the /service directory and load those exploded directory and if someone drop exploded directory as a service that will take into account. 

If the parameter not there or value is false, then none of the archive file will be extracted and if someone drop exploded directory as service that wont take into account. Only load archive files into account. 

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis/trunk/java@264760 13f79535-47bb-0310-9956-ffa450edef68