Completing ServiceGroup issue , changed service descriptor from service.xml to services.xml
Now onward only the valid service descriptor in side a service is services.xml and service.xml wont be a valid one. This is just name change if you only have one service in your service archive but if you have multiple service in it then the services.xml format will be little bit change and the way is as fllows
<serviceGroup>
  <service name=”foo”>
  </service>
 <service name=”bar”>
 </service>
</serviceGroup>

To make the previous  service archive file to be run in the current system I am thinking of providing a way to have either service.xml or services.xml in side META-INF and if both of them are there in META-INF at the deployment time I will only read services.xml and ignore service.xml , and if only the service.xml is there it will work as previous without giving any exception.



git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@289390 13f79535-47bb-0310-9956-ffa450edef68