Patch from Alex and Ralf
This patch does not include any attempts at optimization.  It addresses the
following issues:

1) Resources cannot be loaded from jars inside of the service .aar
getResourceAsStream() has been overridden to allow this.

2) ZipInputStreams were left open by getBytes() if the target was not found.
An additional try/catch block has been added after the search loop to make
sure this happens.

3) Minor change in findLibJars() to normalize the path for startsWith() and
endsWith() checks.


What's missing from this patch is optimization code to address the extended
loading times of the DeploymentClassLoader.  We still don't know how to best
deal with this.

We forgot to extend a big "Thank You" to everyone out there working on this
project in the last post.  Thanks!

--Alex & Ralf
 aartigues@picturemarketing.com
 rsantiago@picturemarketing.com
 


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@351835 13f79535-47bb-0310-9956-ffa450edef68