A fix to get rid of the NPE when we process SOAP 1.1 SOAPFaults using DOOM
Explanation:
faultcode, faultstring, faultactor, detail are not namespace qualified in SOAP 1.1 and SOAP11BuilderHelper tries to do getLocalName on those elements. But if the in this case a DOOM element will return null See: http://java.sun.com/j2se/1.4.2/docs/api/org/w3c/dom/Node.html#getLocalName(). Therefore we have to use getTagName() to get the name of the element if its not qualified



git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@357364 13f79535-47bb-0310-9956-ffa450edef68