Integrate Java2WSDL into deployment section   1. If the WSDL is there then using the WSDL first create an AxisService (thats wt Sanka did)   2. If the wsdl is not there using JAM I generate WSDL
Finally if the WSDL is there or not when you type ?wsdl u should get the wsdl file for that service (which is always DOC-LIT , I will improve that to support any type).

Note : If the service class is not Java or if there are any error at the schema generation time ?wsdl will not work.

Fix Deploymengt Class Loader to support Paranet first and Child first scanario (I have to provide a way to configure that using axis2.xml)

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@358205 13f79535-47bb-0310-9956-ffa450edef68