An itch -> A scratch...
i had some trouble with ADB failing when there is no soap header block, but Axis2 does not let me formulate a message w/o a soap header block. XMLSPY is costly. So what does one do? :)

Adds a panel to HTTP Tracer where you can specify an endpoint, a soap action and the soap message to be sent

Enter the information then click the send and you will see the output in the bottom pane.



git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@375643 13f79535-47bb-0310-9956-ffa450edef68