Added: PolicyEvalutor which evaluates any policy information per operation basis and hands them off to an appropriate PolicyExtension for further processing.
PolicyExtension which processes a set of module specific policy assertions and adds "method" elements to "stubElement" which will be resovled into stub methods. Those stub methods are a function of policy information which is presented in the WSDL and they contirbutes to a generic api which is used to take only the required input from the user. For instance in the username password case it adds only two method (setUsername(), setPassword()) to the stub, not a method like setCallbackClass(). That is because setCallbackClass is internal information required by WSS4J. Hence that information should be generated internally in AxisModule and the user of the stub needs not to provide such information.

CodeGenPolicyExtension is an interface which must be implemeted by Module author who intend to support Policies at codegen time. It contains a single method to retrieve a PolicyExtension for that Module.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@377455 13f79535-47bb-0310-9956-ffa450edef68