Adding a new implementation for the ADBPullParser. This can efficiently handle the null/empty field requirement. Codegen has been changed to reflect the change. 1. Added a ADBXMLStreamReader interface 2. Added implementations for the base ADBStreamReader, NameValueArrayStreamReader (name,values) NameValueArrayStreamReader (name,value[]) and a wrapping reader 3. Changed the StreamWrapper.java to generate an END_DOCUMENT event 4. Changed other classes to suit the change
The most important change in this implementation is that the starting event for the pullparser provided is START_ELEMENT rather than START_DOCUMENT. This makes the reader required to be wrapped in a StreamWrapper when it is needed by the StaxOMBuilder.
However the amount of test cases for ADB does not seem to suffice to completely test the functionality (round trip). The next step should be to implement some roundtrip test cases.



git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@391258 13f79535-47bb-0310-9956-ffa450edef68