interesting piece of code. This was throwing NPEs and contained a return statement inside finally.
Anyway, this lead to me another question. What if the passed value is a null? According to javadoc of the compare method, if the method can not be compared, then one should throw a ClasscastException. I had no option than to throw a ClasscastException even when the passed object is null. 

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@452313 13f79535-47bb-0310-9956-ffa450edef68