Fix for AXIS2-1538 - Axis2 always assumes the web app name to be axis2 when generating the EPR for WSDL ports.
Fighting a losing battle with configContext.getContextRoot()...

1. In ConfigurationContext, there is a hard coded value.
   private String contextRoot = "axis2";
2. Which is of course overrideable by setting contextRoot in axis2.xml

BUT, if there is no override in the axis2.xml in WEB-INF/conf, then we 
get the hardcoded "axis2" value. 

I think i mad a fix at least 1 time (may be more than 1), to override this
default, hard-coded value in ConfigurationContext. 

But the problem is who wins? Should AxisServlet always override the entry
in axis2.xml? probaly not...

So enter a new servlet init param in web.xml named "axis2.find.context",
If it is set to false then we pick up whatever is in axis2.xml/hard-coded 
in ConfigurationContext. If it is true (which is the default case, even
when the param is not specified in web.xml), then we calculate the context.

Hopefully this will not ruffle any feathers! :)




git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/branches/java@468722 13f79535-47bb-0310-9956-ffa450edef68