Added a mechanism to give the current state of the RequestResponseTransport. This could be used by RM in resending of messages. Message will be allowed to be resent only if the RequestResponseTranspor object is in the correct state.
Added a mechanism to allow somebody to set an custom transport response code. 
For e.g. Sandesha2 has a requirement to return HTTP 408, but there was not way do do this in Axis2.

Added a boolean property which should be used by transports to decide weather or not a certain request 
thread should be blocked (by calling RequestResponseTransport.awaitResponse () ) object. Simply doing 
this based on wehater or not the request has been is paused is not enough in some cases.

Please see javadocs for more details.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@489259 13f79535-47bb-0310-9956-ffa450edef68