1. Chnaged signatures of the addContext, removeContext, updateState methods to take an AbstractContext.
This makes the methods more general. IDs can be always obtained by going through the passed contexts correctly.

2. Added a flush() method to the AbstractContexst which calls the clusterManager.updateState() method when clustering is enabled. 

This allows serviceAuthors to write their code in a transparent manner. I.e. if they hv done their state changes with flush() calles they will not have to do any changes to make their services clusterable.

3. Moved the replication point of the execution chain into a Handler.

There is no need to hardcode this into the AxisEngine or MessageReceiver. By placing the ReplicationHandler anyware in the handler chain we can change the replication point as needed. Most probably this will be placed at the end of the flows.
(please see the clustering/conf/axis2.xml).

4. Added a isContextClusterable (AbstractContext) method to the CM.

ClusterManagers can say what contexts they are interested in clustering, this will be useful to the Service authors.

5. Moved the addContext() method calls out of the constructors of the Context classes.

It seems like it's better to have these outside. When we are in the constructor we do not have a fully initiated context object. For e.g. the serviceGroupContextId does not get setted at the time the SGC get created. 
Currently, this made me call clusterManager methods from several places of the core. But I was talking to Deepal about introducing some factories to create Context classes, which will make things much cleaner.

6. Removed the NullClusterManager - as we discussed in the mailing list.

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/branches/java@506412 13f79535-47bb-0310-9956-ffa450edef68