Refactor Modules to be named with Strings, not QNames.
This burbled into an awful lot of places, so I still need to go back and add a few deprecated (commented-out) setters for the QName versions - but I want to check this in so that merging doesn't get outrageous.

Also the usual little bits of code cleanup in here, and remove the explicit locked="false" attributes from the default axis2.xml in the kernel.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@521461 13f79535-47bb-0310-9956-ffa450edef68