Added two changes to the branch
1.
Currently when we send messages out using the HTTP transport, it is 
checked weather a MessageContext object has been set for the 'In' entry 
of the OperationContext. If this is present transport information is 
saved there, else they are ignored. In Sandesha2 we have a scenario where 
this MessageContext is not present in the operation context (the incoming 
message may be a Acknowledgement which should not be added to the OpCtx of 
the application message). So we Axis2 to preserve the transport information 
even when MsgCtx is not set.
This was needed to keep Sandesha2+MTOM working for the 1.2 release

2.
Added a new method named 'signalFaultReady' to the RequestResponseTransport object. 
This will basially ask the Transport to end a current wait throwing the given fault. 





git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/branches/java/1_2@525477 13f79535-47bb-0310-9956-ffa450edef68