Functionality for sending control messages between axis2 nodes in a cluster
At the moment, I've identified 2 control messages;
1. GetStateCommand - Sent by a node which is joining a cluster which already has one or more members. In such a case, it will send a request for the latest
node to a randomly selected member, and will wait till a reply is reveived. If such a reply is not received, it will retry by sending the same request to
another member, or until there are no members in the cluster. Unless a GetStateResponseCommand message is received, a node cannot start successfuly and join a
cluster, unless it is the very first node which is joining the cluster.

2. GetStateResponseCommand - Sent by the node which receives the GetStateCommand. It will populate this message with the latest state and send it out to the
requester.



git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@541316 13f79535-47bb-0310-9956-ffa450edef68