JIRA AXIS2-2943 Contributor: Rich Scheuerle JAX-WS exposes properties to access the inbound and outbound attachments.  The attachments are exposed as a Map<String, DataHandler>.  Axis2 (and the Axis2 JAX-WS implementation) stores attachments on an Axiom AttachmentMap object.
This commit adds a new class, AttachmentAdapter, that is a bridge between the JAX-WS exposed property and the underlying AttachmentMap.  

This commit adds a new class, TransportPropertiesAdapter, which serves a similar function (bridges the JAXWS spec property with the underlying implementation data structure).

This commit updates the JAX-WS handler testcase to test the new adapter classes.  This commit also includes some additional minor changes and logging to the JAX-WS attachments code.

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@556140 13f79535-47bb-0310-9956-ffa450edef68