WSCOMMONS-3581 Contributor: Rich Scheuerle Details of Fix:
1) Added new constructors to WSDL4JWrapper and WSDLDefinitionWrapper to allow the 
caller to pass in the configuration or memory limit parameters.

2) Deprecated the constructors in WSDL4JWrapper and WSDLDefinitionWrapper that
do not pass in configuration or memory limit parameters.  I changed references in the
code to use the new constructors (prefering the ones that pass in the AxisConfiguration).

2a) In the cases where the AxisConfiguration is not available, the code builds a 
WSDL4JWrapper using local information.  For example, if the WSDL4JWrapper is only temporary, then the basic wrapper is used.  For example, if the WSDL4JWrapper is built by JAXWS and
the configuration is not available...then a reload wrapper is sometimes used.

3) Added debug trace information in the WSDLWrapper*Impl constructors that show the call
path for how they were created. 

4) Added debug trace in the Options, AbstractContext and ParameterIncludeImpl classes
to log changes to the properties/parameters.  This can help narrow problems.  (Purely
for support)

Kudos to Vien Tran for working with me on this fix.

git-svn-id: https://svn.apache.org/repos/asf/webservices/axis2/trunk/java@635703 13f79535-47bb-0310-9956-ffa450edef68