JIRA- AXIS2-4845 The change addresses a performance problem in AddressingInHandler. code contributed by Doug Larson.
Currently, we first check to see if there is a parameter to disable inbound addressing which is normally false but then we find out there are no soap headers anyway. The change moves the check for soap headers before we check to see if there is a parameter to disable inbound addressing. 

git-svn-id: https://svn.apache.org/repos/asf/axis/axis2/java/core/trunk@1005915 13f79535-47bb-0310-9956-ffa450edef68