AXIS2-4862 Contributor:Rich Scheuerle Summary: Fixed JAX-WS handler chain processing to propogate the (possibly edited) Message to the Axis2 MessageContext when the handler chain is reversed due to a "return false" from user's JAX-WS handler.   This ensures that the correct message is returned to the inbound handlers and client application.
Also noticed that the JAX-WS fault processing has a similar problem.  Fixed the code in that case too.

Also noticed that the JAX-WS fault processing is not examining the handler's message prior to making a new message.
This contradicts the specification.  I changed to the code to examine the current message and see if it is a fault.
If it is, then that fault is retained.
If not, then the thown protocol Exception is converted into a fault message and becomes the current message.

git-svn-id: https://svn.apache.org/repos/asf/axis/axis2/java/core/trunk@1027138 13f79535-47bb-0310-9956-ffa450edef68