Removing HTML style attributes from tables appearing in Xdoc documents.
Reason: in older versions of the maven-site-plugin, these attributes were simply ignored and had no effect. After upgrading to 2.1.1, they are copied to the XHTML output and thus override the styles defined in the common stylesheet. This is of course unwanted, because it reduces the visual consistency of the site.


git-svn-id: https://svn.apache.org/repos/asf/axis/axis2/java/core/trunk@1039720 13f79535-47bb-0310-9956-ffa450edef68